Use floats rather than integers

+++ run_candidates.rb
@@ -49,22 +49,22 @@ def lat_lon_for_entity(type, id)
   when 'node'
     r = @dbconn.query(NODE_LOCATION_SQL % {:id => id})
     if r.num_tuples > 0
-      lat, lon = r[0]['latitude'].to_i / SCALE, r[0]['longitude'].to_i / SCALE
+      lat, lon = r[0]['latitude'].to_f / SCALE, r[0]['longitude'].to_f / SCALE
     end
   when 'way'
     r = @dbconn.query(WAY_LOCATION_SQL % {:id => id})
     if r.num_tuples > 0
-      lat, lon = r[0]['latitude'].to_i / SCALE, r[0]['longitude'].to_i / SCALE
+      lat, lon = r[0]['latitude'].to_f / SCALE, r[0]['longitude'].to_f / SCALE
     end
   when 'relation'
     r = @dbconn.query(RELATION_LOCATION_BY_NODE % {:id => id})
     if r.num_tuples > 0
-      lat, lon = r[0]['latitude'].to_i / SCALE, r[0]['longitude'].to_i / SCALE
+      lat, lon = r[0]['latitude'].to_f / SCALE, r[0]['longitude'].to_f / SCALE
     else
       puts "2"
       r2 = @dbconn.query(RELATION_LOCATION_BY_WAY % {:id => id})
       if r2.num_tuples > 0
-        lat, lon = r2[0]['latitude'].to_i / SCALE, r2[0]['longitude'].to_i / SCALE
+        lat, lon = r2[0]['latitude'].to_f / SCALE, r2[0]['longitude'].to_f / SCALE
       end
     end
   end
